Fix findbugs

+++ src/main/java/rendering/Renderer.java
@@ -118,8 +118,8 @@ public final class Renderer implements IRenderer {
         String cameraMsg = "Camera corrected Y-position = " + (point.getY() - this.camera.getYPos());
         this.logger.info(drawMsg + cameraMsg);
 
-        double halfWidth = (sprite.getWidth() / 2);
-        double halfHeight = (sprite.getHeight() / 2);
+        double halfWidth = ((double) sprite.getWidth() / 2d);
+        double halfHeight = ((double) sprite.getHeight() / 2d);
         double translateX = point.getX() + halfWidth;
         double translateY = point.getY() - this.camera.getYPos() + halfHeight;
 
@@ -163,8 +163,8 @@ public final class Renderer implements IRenderer {
         String cameraMsg = "Camera corrected Y-position = " + (point.getY() - this.camera.getYPos());
         this.logger.info(drawMsg + cameraMsg);
 
-        double halfWidth = (sprite.getWidth() / 2);
-        double halfHeight = (sprite.getHeight() / 2);
+        double halfWidth = ((double) sprite.getWidth() / 2d);
+        double halfHeight = ((double) sprite.getHeight() / 2d);
         double translateX = point.getX() + halfWidth;
         double translateY = point.getY() - this.camera.getYPos() + halfHeight;
